refactor(studio): split the logs panel's toolbars and summary
LogsPanel was 413 lines, now 291. Four pieces, each a closed set of controls or
one reading of the window:

- `LogsViewBar` — the only chrome common to all three views (view switch, shard,
  live error); the filter bars below it swap per view, this does not.
- `LogsErrorFilters` / `LogsRequestFilters` — per-view controls that previously
  sat adjacent guarded by opposite conditions. `LevelToggle` moved in with the
  filters that render it.
- `LogsSummary` — the aggregate reading, which is the ALTERNATIVE to the entry
  list rather than part of it. Its `LogSummary`/`SummaryBucket` types moved with
  it.

Suppression deleted. 80 logs tests green.

Refs #230.
